# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
...
  /users/favoriteMatches:
    get:
      tags:
        - Users
      summary: Return the favorite matches that were saved by user
      responses:
        "200":
          description: data about the user's favorites matches returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FutureMatchPreview"
        "401":
          description: user is not authorized
        "204":
          description: This user does not have any favorite matches
    post:
      summary: save match in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                match_id:
                  type: int
                  description: match's id which the user saved
                  example: 85
      responses:
        "201":
          description: The match successfully saved as favorite
        "401":
          description: user is not authorized
        "409":
          description: match already in favorites
...
  /search/Teams:
    get:
      tags:
      - Search
      summary: Search for teams by name
      parameters:
      - name: query
        in: query
        description: The query we want to search
        required: true
        style: form
        explode: true
        schema:
          type: string
        example: Midtjylland
      - name: sort
        in: query
        description: type of sort
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
            default: none
            enum:
            - asc
            - desc
            - none
        example: asc
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchTeams'
        "204":
          description: There is no content to send for this request
        "401":
          description: You have not authenticated yet, please sign in
        "404":
          description: The server can not find the requested resource
  /search/Players:
    get:
      tags:
      - Search
      summary: Search for players by name
      parameters:
      - name: query
        in: query
        description: The query we want to search
        required: true
        style: form
        explode: true
        schema:
          type: string
        example: David
      - name: sort
        in: query
        description: type of sort
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
            default: none
            enum:
            - player name
            - team name
            - none
        example: player name
      - name: filter
        in: query
        description: how to filter results, can be done by position of player or by team name
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
            default: none
            enum:
            - team name
            - player's position
            - none
        example: player's position
      - name: filter query
        in: query
        description: which value do you want to be in the filter
        required: false
        style: form
        explode: true
        schema:
          type: string
        example: 3
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerPreview'
        "204":
          description: There is no content to send for this request
        "401":
          description: You have not authenticated yet, please sign in
        "404":
          description: The server can not find the requested resource
...
components:
  schemas:
...
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust
        id:
          type: integer
          description: 172923
          example: Sten Michael Grytebust
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
        position:
          type: integer
          description: player's position in team
          example: 1
        team_name:
          type: string
          description: player's team name
          example: København
...
    FutureMatchPreview:
      type: object
      properties:
        match_id:
          type: integer
          example: 53
        date_time:
          type: string
          format: date-time
          example: "2020-08-02 15:30:00"
        local_team_id:
          type: integer
          example: 53
        visitor_team_id:
          type: integer
          example: 338
        venue_id:
          type: integer
          example: 8909 
...
    SearchTeams:
      type: object
      properties:
        team name:
          type: string
          example: Midtjylland
        logo path:
          type: string
          format: URL
          example: https://cdn.sportmonks.com/images//soccer/teams/11/939.png